@@ -8,7 +8,7 @@
     ],
     "license": "MIT",
     "require": {
-        "php": "^7.1.3",
+        "php": "^7.4|^8.0",
         "fernandowobeto/model-fillable-control": "1.1",
         "fernandowobeto/uol-motor1-rss": "1.1",
         "fideloper/proxy": "^4.0",
